refactor(js): change controller implementation
Instead of using multiple traits to expose functionality for the RPC
and Join service respectively, the controller exposes a browser and
context separately, leading to a massively simplified implementation.

The downside is that there is less abstraction, and e.g the RPC service
can add clients to the context, but the downside is definitely worth it.